Move text list files to module
