+++ app/src/main/java/ir/mehdiyari/krypt/ui/navigation/KryptNavHost.kt
@@ -8,7 +8,6 @@ import ir.mehdiyari.krypt.features.auth.login.navigateToLogin
 import ir.mehdiyari.krypt.setting.ui.settingsRoute
 import ir.mehdiyari.krypt.ui.KryptAppState
 import ir.mehdiyari.krypt.ui.data.dataScreen
-import ir.mehdiyari.krypt.shareContent.ShareDataViewModel
 import ir.mehdiyari.krypt.ui.home.homeScreen
 import ir.mehdiyari.krypt.ui.home.navigateToHome
 import ir.mehdiyari.krypt.features.auth.create_account.createAccountScreen
@@ -17,8 +16,8 @@ import ir.mehdiyari.krypt.ui.media.mediaScreen
 import ir.mehdiyari.krypt.ui.media.navigateToMedia
 import ir.mehdiyari.krypt.ui.text.add.addTextScreen
 import ir.mehdiyari.krypt.ui.text.add.navigateToAddText
-import ir.mehdiyari.krypt.ui.text.list.navigateToTexts
-import ir.mehdiyari.krypt.ui.text.list.textsScreen
+import ir.mehdiyari.krypt.features.text.list.navigateToTexts
+import ir.mehdiyari.krypt.features.text.list.textsScreen
 import ir.mehdiyari.krypt.ui.voice.audios.audiosRoute
 import ir.mehdiyari.krypt.ui.voice.audios.navigateToAudios
 import ir.mehdiyari.krypt.ui.voice.record.addVoiceScreen
