+++ app/src/main/java/ir/mehdiyari/krypt/ui/text/add/AddTextViewModel.kt
@@ -7,10 +7,10 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import ir.mehdiyari.krypt.R
 import ir.mehdiyari.krypt.dispatchers.di.DispatchersQualifierType
 import ir.mehdiyari.krypt.dispatchers.di.DispatchersType
+import ir.mehdiyari.krypt.features.text.logic.TextEntity
 import ir.mehdiyari.krypt.file.data.entity.FileEntity
 import ir.mehdiyari.krypt.file.data.entity.FileTypeEnum
 import ir.mehdiyari.krypt.files.logic.repositories.api.FilesRepository
-import ir.mehdiyari.krypt.ui.text.list.TextEntity
 import ir.mehdiyari.krypt.features.text.logic.TextFilesUtils
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -136,7 +136,7 @@ class AddTextViewModel @Inject constructor(
                 if (titleContentPair != null) {
                     _argsTextViewState.emit(
                         AddTextArgsViewState.TextArg(
-                            TextEntity(
+                           TextEntity(
                                 textId,
                                 titleContentPair.first,
                                 titleContentPair.second
