+++ features/text/list/build.gradle.kts
@@ -14,4 +14,15 @@ android {
 
 dependencies {
 
+    implementation(project(":shared:designsystem:theme"))
+    implementation(project(":shared:designsystem:resources"))
+
+    implementation(project(":shared:dispatchers"))
+    implementation(project(":shared:files:files-logic"))
+    implementation(project(":shared:files:files-data"))
+    implementation(project(":features:text:shared"))
+
+    implementation(libs.androidx.lifecycle.viewmodelCompose)
+    implementation(libs.androidx.lifecycle.runtimeCompose)
+    implementation(libs.androidx.hilt.navigation.compose)
 }
