+++ features/text/list/src/main/res/drawable/ic_pen.xml
