+++ features/text/shared/.gitignore
