+++ features/text/shared/src/main/AndroidManifest.xml
