+++ features/text/shared/src/main/java/ir/mehdiyari/krypt/features/text/logic/TextFilesUtils.kt
