+++ settings.gradle.kts
@@ -34,5 +34,5 @@ include(":features:settings")
 include(":shared:files:backup-logic")
 include(":features:auth:create-account")
 include(":shared:share-content")
-include(":features:text:logic")
+include(":features:text:shared")
 include(":features:text:list")
