+++ shared/designsystem/resources/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
     <string name="YES">Yes</string>
     <string name="menu_settings">Settings</string>
     <string name="password_not_match">Passwords does not match</string>
+    <string name="texts_library">Texts</string>
 </resources>
